--- src/components/todo/Todo.scss
@@ -1,18 +1,0 @@
-.todo {
-    display: flex;
-
-    &.hidden {
-        max-height: 100vh;
-    }
-
-    .todo-main {
-        flex-grow: 1;
-        max-width: 1300px;
-        margin: 0 auto;
-        padding: 0 50px;
-
-        &.hidden {
-            overflow-x: hidden;
-        }
-    }
-}
