+++ src/modules/Firestore/selectors/index.js
@@ -11,6 +11,7 @@ export const selectFirestoreData = createSelector(
   (state) => state.data
 )
 
+// deleted todos
 export const selectDeletedTodos = createSelector(
   selectFirestoreData,
   (state) => state.deletedTodos && Object.entries(state.deletedTodos).map(([key, value]) => ({
@@ -23,3 +24,35 @@ export const selectRequestingDeletedTodos = createSelector(
   selectFirestore,
   (state) => state.status.requesting.deletedTodos
 )
+
+// categories
+export const selectCategories = createSelector(
+  selectFirestoreData,
+  (state) => state.categories && Object.entries(state.categories).map(([key, value]) => (value && {
+    id: key,
+    ...value
+  }))
+)
+
+export const selectRequestingCategories = createSelector(
+  selectFirestore,
+  (state) => state.status.requesting.categories
+)
+
+// subcategories
+export const selectSubCategories = (state, categoryID) => {
+  return createSelector(
+    selectFirestoreData,
+    (state) => state[`subcategory-${categoryID}`] && Object.entries(state[`subcategory-${categoryID}`]).map(([key, value]) => ( value && {
+      id: key,
+      ...value
+    }))
+  )(state)
+}
+
+export const selectRequestingSubCategories = (state, categoryID) => {
+  return createSelector(
+    selectFirestore,
+    (state) => state.status.requesting[`subcategory-${categoryID}`]
+  )(state)
+}
