+++ src/pages/Todo/styles.scss
@@ -1,0 +1,18 @@
+.todo {
+  display: flex;
+
+  &.hidden {
+      max-height: 100vh;
+  }
+
+  .todo-main {
+      flex-grow: 1;
+      max-width: 1300px;
+      margin: 0 auto;
+      padding: 0 50px;
+
+      &.hidden {
+          overflow-x: hidden;
+      }
+  }
+}
